Split CR e2e tests into postgres and bigquery suites
Move existing component readiness e2e tests under postgres/ and restore
the upstream BigQuery TestComponentReadinessViews test in bigquery/,
gated on GCS_SA_JSON_PATH. The e2e script now runs two phases: postgres
tests first, then restarts the server with --data-provider bigquery for
the BQ suite when credentials are available.
