+++ scripts/e2e.sh
@@ -12,12 +12,16 @@ REDIS_CONTAINER="sippy-e2e-test-redis"
 REDIS_PORT="23479"
 
 if [ -z "$GCS_SA_JSON_PATH" ]; then
-    echo "WARNING: GCS_SA_JSON_PATH not set, data sync test will be skipped" 1>&2
+    echo "WARNING: GCS_SA_JSON_PATH not set, data sync and BigQuery tests will be skipped" 1>&2
 fi
 
+E2E_EXIT_CODE=0
 
 clean_up () {
     ARG=$?
+    if [ $ARG -ne 0 ]; then
+        E2E_EXIT_CODE=$ARG
+    fi
     echo "Stopping sippy API child process: $CHILD_PID"
     kill $CHILD_PID 2>/dev/null && wait $CHILD_PID 2>/dev/null
     # Generate coverage report from the server's coverage data
@@ -26,11 +30,13 @@ clean_up () {
         go tool covdata percent -i="$COVDIR"
         go tool covdata textfmt -i="$COVDIR" -o=e2e-coverage.out
         # Merge test binary coverage (from -coverprofile) into server binary coverage
-        if [ -f e2e-test-coverage.out ]; then
-            echo "Merging test binary coverage into server coverage..."
-            tail -n +2 e2e-test-coverage.out >> e2e-coverage.out
-            rm -f e2e-test-coverage.out
-        fi
+        for f in e2e-test-coverage.out e2e-bq-test-coverage.out; do
+            if [ -f "$f" ]; then
+                echo "Merging $f into server coverage..."
+                tail -n +2 "$f" >> e2e-coverage.out
+                rm -f "$f"
+            fi
+        done
         echo "Coverage data written to e2e-coverage.out"
         echo "View HTML report: go tool cover -html=e2e-coverage.out -o=e2e-coverage.html"
     fi
@@ -40,7 +46,23 @@ clean_up () {
     echo "Tearing down container $REDIS_CONTAINER"
     $DOCKER stop -i $REDIS_CONTAINER
     $DOCKER rm -i $REDIS_CONTAINER
-    exit $ARG
+    exit $E2E_EXIT_CODE
+}
+
+wait_for_sippy() {
+    echo "Waiting for sippy API to start on port $SIPPY_API_PORT..."
+    TIMEOUT=600
+    ELAPSED=0
+    while [ $ELAPSED -lt $TIMEOUT ]; do
+        if curl -s "http://localhost:$SIPPY_API_PORT/api/health" > /dev/null 2>&1; then
+            echo "Sippy API is ready after ${ELAPSED}s"
+            return 0
+        fi
+        sleep 2
+        ELAPSED=$((ELAPSED + 2))
+    done
+    echo "Timeout waiting for sippy API to start after ${TIMEOUT}s"
+    return 1
 }
 trap clean_up EXIT
 
@@ -95,23 +117,7 @@ GOCOVERDIR="$COVDIR" ./sippy serve \
   --data-provider postgres > e2e.log 2>&1 &
 CHILD_PID=$!
 
-# Give it time to start up, and fill the redis cache
-echo "Waiting for sippy API to start on port $SIPPY_API_PORT, see e2e.log for output..."
-TIMEOUT=600
-ELAPSED=0
-while [ $ELAPSED -lt $TIMEOUT ]; do
-    if curl -s "http://localhost:$SIPPY_API_PORT/api/health" > /dev/null 2>&1; then
-        echo "Sippy API is ready after ${ELAPSED}s"
-        break
-    fi
-    sleep 2
-    ELAPSED=$((ELAPSED + 2))
-done
-
-if [ $ELAPSED -ge $TIMEOUT ]; then
-    echo "Timeout waiting for sippy API to start after ${TIMEOUT}s"
-    exit 1
-fi
+wait_for_sippy || exit 1
 
 # Prime the component readiness cache so triage tests can find cached reports
 echo "Priming component readiness cache..."
@@ -122,7 +128,45 @@ for VIEW in $(echo "$VIEWS" | jq -r '.[].name'); do
 done
 echo "Cache priming complete"
 
-# Run our tests that request against the API, args ensure serially and fresh test code compile:
-gotestsum ./test/e2e/... -count 1 -p 1 -coverprofile=e2e-test-coverage.out -coverpkg=./pkg/...,./cmd/...
+# Phase 1: Run postgres-backed tests
+echo "=== Phase 1: Running postgres-backed e2e tests ==="
+gotestsum \
+  ./test/e2e/componentreadiness/postgres/... \
+  ./test/e2e/componentreadiness/bugs/... \
+  ./test/e2e/datasync/... \
+  ./test/e2e/ \
+  -count 1 -p 1 -coverprofile=e2e-test-coverage.out -coverpkg=./pkg/...,./cmd/...
+POSTGRES_EXIT=$?
+if [ $POSTGRES_EXIT -ne 0 ]; then
+    E2E_EXIT_CODE=$POSTGRES_EXIT
+fi
 
-# WARNING: do not place more commands here without addressing return code from go test not being overridden by the cleanup func
+# Phase 2: Run BigQuery-backed tests (if credentials are available)
+if [ -n "$GCS_SA_JSON_PATH" ]; then
+    echo "=== Phase 2: Running BigQuery-backed e2e tests ==="
+    echo "Stopping postgres-backed server..."
+    kill $CHILD_PID 2>/dev/null && wait $CHILD_PID 2>/dev/null
+
+    GOCOVERDIR="$COVDIR" ./sippy serve \
+      --listen ":$SIPPY_API_PORT" \
+      --listen-metrics ":12112" \
+      --database-dsn="$SIPPY_E2E_DSN" \
+      --log-level debug \
+      --views config/e2e-views.yaml \
+      --google-service-account-credential-file $GCS_SA_JSON_PATH \
+      --redis-url="$REDIS_URL" \
+      --data-provider bigquery > e2e-bq.log 2>&1 &
+    CHILD_PID=$!
+
+    wait_for_sippy || exit 1
+
+    gotestsum \
+      ./test/e2e/componentreadiness/bigquery/... \
+      -count 1 -p 1 -coverprofile=e2e-bq-test-coverage.out -coverpkg=./pkg/...,./cmd/...
+    BQ_EXIT=$?
+    if [ $BQ_EXIT -ne 0 ]; then
+        E2E_EXIT_CODE=$BQ_EXIT
+    fi
+else
+    echo "=== Phase 2: Skipping BigQuery tests (GCS_SA_JSON_PATH not set) ==="
+fi
