+++ test/e2e/componentreadiness/postgres/componentreadiness_test.go
@@ -1,4 +1,4 @@
-package componentreadiness
+package postgres
 
 import (
 	"fmt"
